Use service instead of dnsRecord
