+++ lib/shared/addon/all-dns-records/service.js
@@ -29,7 +29,7 @@ export default Service.extend({
     this._super(...arguments);
     const store = get(this, 'store');
 
-    set(this, '_allDnsRecords', store.all('dnsRecord'));
+    set(this, '_allDnsRecords', store.all('service'));
   },
 
   dnsRecords: function() {
