+++ lib/shared/addon/components/containers-header/template.hbs
@@ -6,7 +6,7 @@
     {{#if (rbac-allows resource="ingress" scope="project" permission="list")}}
       <li>{{#link-to-external "ingresses.index" scope.currentProject.id}}{{t 'nav.containers.ingresses'}}{{/link-to-external}}</li>
     {{/if}}
-    {{#if (rbac-allows resource="dnsrecord" scope="project" permission="list")}}
+    {{#if (rbac-allows resource="service" scope="project" permission="list")}}
       <li>{{#link-to-external "authenticated.project.dns.index" scope.currentProject.id}}{{t 'nav.containers.dns'}}{{/link-to-external}}</li>
     {{/if}}
     <li>{{#link-to-external "volumes.index" scope.currentProject.id}}{{t 'nav.containers.volumes'}}{{/link-to-external}}</li>
